+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased] 
+
+### Added
+
+- [TD-2586] Fields `last_execution` and `execution_status` on search 
+
 ## [3.19.0] 2020-04-06
 
 ### Added