+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- [TD-2490] td-cache: Updated Taxonomy information on domain update
+
 ## [3.19.0] 2020-04-06
 
 ### Added